% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dq_ip_file}
\alias{dq_ip_file}
\title{IP File DQ Assessment}
\format{
A data frame with five variables:
\itemize{
\item \code{State}: State name
\item \verb{Data Year}: TAF data year
\item \code{Variable}: TAF variable name
\item \code{Topic}: Data quality topic
\item \verb{DQ Assessment}: Data quality assessment conducted by the DQ Atlas project
}
}
\source{
\url{https://www.medicaid.gov/dq-atlas/landing/topics/info}
}
\usage{
dq_ip_file
}
\description{
DQ Assessment at the state-year-variable level for the TAF IP file
}
\keyword{datasets}
