% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqtools-package.R
\docType{package}
\name{dqtools-package}
\alias{dqtools}
\alias{dqtools-package}
\title{dqtools: Data Quality Assessment from DA Atlas}
\description{
`dqtools` allows users to quickly view the data quality assessments for TAF variables.
}
\author{
\strong{Maintainer}: JJ Chen \email{jiajia.chern@gmail.com} (\href{https://orcid.org/0000-0001-8482-8398}{ORCID})

}
\keyword{internal}
